Variance round 1
The variance detection clearly needs some calibration. However, it is at least finding changes in variance, and sometimes even getting the line right (albeit the fact that it is telling me that a variance change on line 60 of a file that in fact has a mean change on line 70 is a little suspicious). I also made it so the program prints what kind of change it finds (variance/mean) at a line. 

There is also something called getLine. Theoretically getLine could more accurately get the line where the variance changed. Something similar could be used for getting the line where the median changes. My understanding right now is that we just get the line at the beginning of the window and call that the line that the change occurred.